Added BT_GetMachineDescription API, for initial boot stages to get the machine descriptor.
This allows a platform independent boot sequence, where all required HAL interfaces can be
found in a single MACHINE_DESCRIPTION structure.

In future we could add the ability for the kernel to support multiple machine descriptions,
and have the bootload pass us a machine/id or machine boot string which we can grep.

This would allow a single kernel with multiple HALs to be built, covering multiple SUBARCH systems.
Obviously they'd all have to be the same architecture. In anycase it could solve many complex management
issue where we have multiple systems with differing variations of PCBs etc, where we want to keep
some boards back on a particular HAL revision.

This must work, because by definition any HAL which is not completely compliant with the HAL interface
specification is broken..