Added a really simple (fragile) heap algorithm.
This heap implementation absorbs the heap space provided by the linker script.
It should really be only used at the kernel-space level for now.

When we have a properly designed memory manager with page management e.g. for devices with
MMUs and MPUs then we do some much nicer stuff.

In anycase systems with all the complexity configured away will resort back to this
heap implementation, or a variation of it.

IDEAS:
      B+Tree HEAP, O(log n) lookup of freespace pockets.

I mention this is fragile, the main reason is that there's basically no boundary between the allocated
memory, and the heap structure data. Off-by-one errors could easily cause HEAP corruption.

Later we'll abstract this with memory pools, and internal heap tables.