[API] High and Low-resolution timestamping now available.
The current implementation is quite crude, and relies kernel ticks being 1ms.
However this should be improved later to satisfy the following conditions:

	* No use of division.
	* Use of any kernel tick rate.

The low-resolution time-stamp is simply a kernel tick count, whereas the
BT_GetKernelTime API actually returns time in µS since the kernel was started.

Note... THIS IS NOT THE TIME SINCE THE SYSTEM STARTED.