[OS] DevFS - Added DEVFS_INODE structure, define macro, and BT_DeviceOpen() api.
This allows machin descriptors to add an entry into the /dev/ folder on the VFS.
Furthermore, the items relate to BT_INTEGRATED_DEVICE structures, and so they
simple call device driver probe functions.

The nice thing is, is that all drivers look identical whether they are managed
or simply exported through the filesystem.

Also for really small systems, this technique requires no RAM, and only a little
extra ROM for each added entry.

The .bt.module.entries linker section is now deprecated, and replaced with:
.bt.devfs.entries.